Moved ExtractSummary tests to new file
