+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/SessionRecords/ExtractSummaryTests/ExtractSummaryBatchWithErrorTest.json
